pyproject.toml
[project]
name = "rpa-anuencia"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.38.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)"
]
